Refactored the NamesSorter
I noticed I was parsing in the entire IClient interface where the NamesSorter only cared about and required the ICommandLineOutputSender interface. This allows the code to be more segregated and modular, as it only needs access to the error message sending functionality. Now if another implementation of this class were to be used, it would only require the sending interface not the entire client to be implemented.